Fixes for 0001-CVE-2020-9327, 0003-CVE-2020-11655 and 0005-CVE-2020-11656
Now we have a way to build sqlite from actual sources which means we
can patch them.

Please see the pathset jinja2 variable at the top of the file.

Although it looks a lot more complicated, if you use patchset=amalgamated
then the actual rendered recipe is just as simple as the old one!

Also, we swap now drop readline and use libedit instead.

Finally, the also fixes expose_symbols.patch so it works on other OSes too
and does not need to be skipped.